+++ packages/mongo/polling_observe_driver.js
@@ -75,7 +75,7 @@ PollingObserveDriver = function (options) {
   // Make sure we actually poll soon!
   self._unthrottledEnsurePollIsScheduled();
 
-  Package.facts && Package.facts.Facts.incrementServerFact(
+  Package['facts-base'] && Package['facts-base'].Facts.incrementServerFact(
     "mongo-livedata", "observe-drivers-polling", 1);
 };
 
@@ -213,7 +213,7 @@ _.extend(PollingObserveDriver.prototype, {
     _.each(self._pendingWrites, function (w) {
       w.committed();
     });
-    Package.facts && Package.facts.Facts.incrementServerFact(
+    Package['facts-base'] && Package['facts-base'].Facts.incrementServerFact(
       "mongo-livedata", "observe-drivers-polling", -1);
   }
 });